feat: ssp-api代码增加

+++ open-adexchange-dao/src/main/java/top/openadexchange/dao/AdPlacementDao.java
@@ -12,9 +12,9 @@ import top.openadexchange.model.AdPlacement;
  */
 public interface AdPlacementDao extends IService<AdPlacement> {
 
-    Boolean enableAdPlacement(Long id);
+    Boolean enableAdPlacement(Integer id);
 
-    Boolean disableAdPlacement(Long id);
+    Boolean disableAdPlacement(Integer id);
 
     AdPlacement getByTemplateCode(String nativeTemplateCode);
 }
